pyproject.toml
[project]
name = "dbt-bq-sourcegen"
version = "0.1.0"
description = "Create or update BigQuery source YAML for dbt"
readme = "README.md"
authors = [{ name = "K-Oxon", email = "ko1011qfp@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "google-cloud-bigquery>=3.34.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "ruamel-yaml>=0.18.14",
]

[project.scripts]
dbt-bq-sourcegen = "dbt_bq_sourcegen.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dbt_bq_sourcegen"]

[tool.hatch.version]
path = "src/dbt_bq_sourcegen/__init__.py"

[dependency-groups]
dev = [
    "ruff>=0.12.2",
 "pytest>=7.0",
 "pytest-cov>=4.0",
 "pyright>=1.1.402",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
